requirements.txt
asgiref==3.9.1
#certifi==2025.8.3
#charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
#idna==3.10
mysqlclient==2.2.7
#PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
#urllib3==2.5.0
Pillow 11.3.0
gunicorn   #manuel
whitenoise  #manuel
psycopg2-binary  #manuel
dj-database-url  #manuel